Portability and rendering improvements
Decoupled encode/decode code from dependence on Paint.NET, meaning the actual 'guts' of the filter are now much more portable. Images are now scaled to a fixed size (width in addition to height) before being processed to allow small images to be filtered correctly, and to reduce the time it takes to convert large images.